requirements.txt
django
djangorestframework
django-environ
djangorestframework-simplejwt